Changes in Android Launcher to support obfuscated fonts
These changes piggyback on the changes recently checked in
the feature/AndroidByteRangeStream branch for Radium-SDK.
In that set of changes, we modified the code in Readium-SDK to
support obfuscated fonts in Android. We did so by modifying the
code in the JNI layer to make use of Package::GetFilterChainByteStream().
However, this required other changes, which modified the interface for
the JNI layer and as such required modifications in the Launcher itself.
This checkin contains those changes.