Support for byte ranges in the Android Launcher
This checkin includes some changes that will allow the Android launcher
to support byte ranges. In a nutshell, the Package class from Readium SDK
will return a ByteBuffer to the EpubServer, when a range request has been made.
The EpubServer code will get that ByteBuffer and put it into a
ByteArrayInputStream, and that will be the InputStream that will be returned to
the Nano HTTP Server.