Adding code to deal with ContentFilter errors in the Launcher
This is part of a larger work item to support errors in ContentFilters.
The ContentFilters are the basic mechanism for extension of the
Readium SDK. As such, people all over the world will start writing
their own ContentFilters, and those ContentFilters will start having
errors. I added a mechanism for reporting those errors, but the
Android Launcher wasn't listenting to those errors. I modified the
Launcher to listen to those errors and then log them. Notice that this
is just to provide to developers with sample code about how to listen
to ContentFilter errors so that they can write code to deal with those
errors properly.